+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 2017-01-21 Chase Hennion
+
+* Adding work that covers lecture 46 through 48 in Section 2.
+* Adding isograms.txt, a file that contains a variety of 3, 4, 5, 6 and 7 letter isograms.
+* Adding handling to game to ask player for desired isogram length
+* Adding handling to game to read isograms from isograms.txt and select one at random for each game based upon the desired isogram length.
+* Adding a neat ASCII banner to the intro.
+
 ## 2017-01-20 Chase Hennion
 
 * Adding work that covers lectures 37 through 45 in Section 2.